Fix pong paddle bug in 8XY5!
I should pay more attention to documentation and this is a great time to learn that unit testing isn't perfect if you assert the wrong result :facepalm:. By stepping through with the new debugger and reading carefully I found that 8XY5 was setting VF to 1 when there was a borrow, 0 otherwise; this behaviour should be the reverse: 0 when borrow, 1 otherwise. Pong seems to now run perfectly and that pesky collision detection issue is gone.